# Gollek Native Build Stabilization — Complete
## Result: ✅ BUILD SUCCESS

**Binary**: `ui/gollek-cli/target/gollek` — **293MB Mach-O arm64** executable
**Build time**: ~5 min 11s
**Installed at**: `~/.local/bin/gollek`

---

## Root Cause Analysis

The native-image build was blocked by **two distinct failure categories**:

### 1. CDI API Version Mismatch (Fatal)

Quarkus 3.32.2 requires **CDI 4.1** (`jakarta.enterprise.invoke.Invoker`), but the project had explicit version pins to **CDI 4.0.1** in two places:

| File | Line | Impact |
|------|------|--------|
| [pom.xml](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/gollek/pom.xml) | ~86 | Parent `dependencyManagement` override |
| [bom/pom.xml](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/gollek/bom/pom.xml) | ~62 | gollek-bom `dependencyManagement` override |

Both overrode the Quarkus BOM's CDI 4.1.0, causing `VertxEventBusConsumerRecorder` to fail at link time with `Discovered unresolved type: jakarta.enterprise.invoke.Invoker`.

**Fix**: Removed all three Jakarta EE version pins (CDI, inject, validation) and let the Quarkus BOM manage them. Also removed the explicit `jakarta.enterprise.cdi-api` dependency from `gollek-cli/pom.xml` since `quarkus-arc` provides it transitively.

### 2. GraalVM Image Heap Initialization (Iterative)

Multiple `_Bean` classes and package-level `--initialize-at-run-time` rules conflicted with Quarkus's CDI container being baked into the image heap:

- `SafetensorShardLoader_Bean`, `SafetensorLoadCache_Bean`, `SafetensorValidator_Bean`, etc.
- Package-level rules for `tech.kayys.gollek.safetensor.loader`, `tech.kayys.gollek.converter.java`, etc.

**Fix**: Removed all `_Bean`-level and overly broad package-level `--initialize-at-run-time` rules. Only kept narrow rules for classes that genuinely instantiate `Cleaner` or native resources at `<clinit>`:

- `org.jline.nativ.Kernel32` (and inner classes)
- `tech.kayys.gollek.inference.libtorch.core.TorchTensor` / `$CleanerHolder`
- `tech.kayys.gollek.safetensor.core.tensor.AccelTensor`
- `tech.kayys.gollek.llama` / `tech.kayys.gollek.inference.llamacpp`

---

## Final `application.properties` Native Args

```properties
quarkus.native.additional-build-args[0]=--enable-native-access=ALL-UNNAMED
quarkus.native.additional-build-args[1]=-H:+ReportExceptionStackTraces
quarkus.native.additional-build-args[2]=-H:+AddAllCharsets
quarkus.native.additional-build-args[3]=-H:+UnlockExperimentalVMOptions
quarkus.native.additional-build-args[4]=-H:+VectorAPISupport
quarkus.native.additional-build-args[5..15]=--initialize-at-run-time=org.jline.nativ.Kernel32*
quarkus.native.additional-build-args[16..17]=--initialize-at-run-time=TorchTensor*
quarkus.native.additional-build-args[18]=-H:-SharedArenaSupport
quarkus.native.additional-build-args[19..20]=--initialize-at-run-time=gollek.llama,gollek.inference.llamacpp
quarkus.native.additional-build-args[21]=--initialize-at-run-time=AccelTensor
quarkus.native.additional-build-args[22..25]=--enable-preview, --add-modules=jdk.incubator.vector
quarkus.native.additional-build-args[26]=--trace-object-instantiation=java.lang.ref.Cleaner
```

> [!IMPORTANT]
> **Key lesson**: Never pin Jakarta EE API versions in `dependencyManagement` when using Quarkus — the Quarkus BOM must be the sole authority for these versions.

> [!WARNING]
> The `--initialize-at-run-time` flag for package-level rules (e.g. `tech.kayys.gollek.safetensor.loader`) will also catch Quarkus-generated `_Bean` classes in that package, causing circular conflicts. Always use class-level rules targeting only the specific classes that instantiate `Cleaner`/`Arena`.

+++ framework/lib/gollek-ml-core/src/main/java/tech/kayys/gollek/ml/calibration/CalibratedClassifierCV.java
@@ -1,7 +1,8 @@
 package tech.kayys.gollek.ml.calibration;
 
 import tech.kayys.gollek.ml.base.*;
-import tech.kayys.gollek.ml.model_selection.*;
+import tech.kayys.gollek.ml.model_selection.ModelSelection.KFold;
+import tech.kayys.gollek.ml.model_selection.ModelSelection.Fold;
 import java.util.*;
 
 /**
@@ -222,8 +223,8 @@ public class CalibratedClassifierCV extends BaseEstimator {
             // Generate points for interpolation
             points = new ArrayList<>();
             for (Block block : blocks) {
-                points.add(new Point(block.xStart, block.mean));
-                points.add(new Point(block.xEnd, block.mean));
+                points.add(new Point((float) block.xStart, block.mean));
+                points.add(new Point((float) block.xEnd, block.mean));
             }
         }
